requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.3
attrs==23.1.0
babelfish==0.6.0
charset-normalizer==3.3.0
cryptg==0.4.0
frozenlist==1.4.0
greenlet==3.0.0
guessit==3.7.1
idna==3.4
multidict==6.0.4
numpy==1.26.0
pandas==2.1.1
pyaes==1.6.1
pyasn1==0.5.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
query==0.1.4
rebulk==3.2.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.21
Telethon==1.30.3
typing_extensions==4.8.0
tzdata==2023.3
yarl==1.9.2
